backend/requirements.txt
annotated-types==0.7.0
bidict==0.23.1
blinker==1.9.0
CacheControl==0.14.2
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
crcmod==1.7
decorator==5.1.1
Deprecated==1.2.18
dnspython==2.7.0
docstring_parser==0.16
eventlet==0.39.0
firebase-admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-aiplatform==1.79.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-pubsub==2.28.0
google-cloud-resource-manager==1.14.0
google-cloud-scheduler==2.15.1
google-cloud-secret-manager==2.22.1
google-cloud-speech==2.30.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.24.0
google-cloud-translate==3.19.0
google-cloud-vision==3.9.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.6.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
moviepy==2.1.2
msgpack==1.1.0
msgspec==0.19.0
numpy==2.2.2
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pillow==10.4.0
proglog==0.1.10
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==24.2.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
rsa==4.9
shapely==2.0.6
simple-websocket==1.1.0
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
